8377172: Change datatype of CodeEntryAlignment to uint
The main idea of this change is to remove an implicitly narrowing
conversion to int from intx (`int code_alignment()`).

CodeEntryAlignment is constrained to not be bigger than
CodeCacheSegmentSize that is itself constrained to be in the range(1,
1024). The data type should really be a uint16_t as its allowed range
is [16-1024], but we have no such support for VM flags. uint will do,
and will simplify the code greatly (five less casts, six if we count
the implicit narrowing cast).

Also add explicit range on the flag.

+++ src/hotspot/share/opto/constantTable.cpp
@@ -158,7 +158,7 @@ void ConstantTable::calculate_offsets_and_size() {
   // Align size up to the next section start (which is insts; see
   // CodeBuffer::align_at_start).
   assert(_size == -1, "already set?");
-  _size = align_up(offset, (int)CodeEntryAlignment);
+  _size = align_up(offset, CodeEntryAlignment);
 }
 
 bool ConstantTable::emit(C2_MacroAssembler* masm) const {
